requirements.txt
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.3
django-embed-video==1.4.3
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
h11==0.12.0
httpcore==0.13.7
httplib2==0.20.4
httpx==0.19.0
idna==3.3
messages==0.8.0
numpy==1.22.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
python-decouple==3.6
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
validus==0.3.0
